fix the design
