+++ package.json
@@ -14,6 +14,7 @@
     "@react-navigation/native-stack": "^7.3.14",
     "react": "19.0.0",
     "react-native": "0.79.3",
+    "react-native-linear-gradient": "^2.8.3",
     "react-native-safe-area-context": "^5.4.1",
     "react-native-screens": "^4.11.1",
     "react-native-svg": "^15.12.0"
