+++ src/constants/colors.ts
@@ -6,5 +6,15 @@ export const colors = {
   //   text colors
   primaryText: '#000000',
   secondaryText: '#494949',
-  userNmaeText:'#D12765',
+  userNmaeText: '#D12765',
+
+  //shadow
+
+  glow_shadow: '#FFE038',
+
+  //surface color
+  primary_surface: '#fff',
+  secondary_surface: '#FFEBEC',
+  secondary_surface_2: '#FCEEED',
+  tertiary_surface: '#FFF2DE',
 };
